+++ src/mongo-wire.c
@@ -112,7 +112,7 @@ mongo_wire_packet_set_header (mongo_packet *p,
       errno = EINVAL;
       return FALSE;
     }
-  if (header->length < sizeof (mongo_packet_header))
+  if (header->length < (gint32)sizeof (mongo_packet_header))
     {
       errno = ERANGE;
       return FALSE;
